feat: Refactor SboVec class and import it in relevant files
